+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 ### Added
 - Treat `BaseElement` as a simple container, omitting the start and end tags when
 rendering.
+### Removed
+- `BaseElement.children` and `BaseElement.attributes` are no longer public.
 
 ## [0.1.0] - 2021-11-08
 - Initial release